+++ src/main/java/edu/yu/marketmaker/model/ReservationResponse.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public record ReservationResponse(
         UUID id,
         ReservationStatus status,
-        long grantedQuantity
+        int grantedQuantity
 )
 {
